+++ RootsFinderSquareSolver.cpp
@@ -2,15 +2,16 @@
 #include <cmath>
 #include <limits.h>
 #include <assert.h>
+
 #include "RootsFinderSquareSolver.h"
 #include "SubsidiaryFunctionsSquareSolver.h"
 
-int find_result(float a, float b, float c, float *result1, float *result2);
-float calculate_discriminant(float a, float b, float c);
-int find_linear_root(float b, float c, float *result1);
-int find_quadratic_roots(float a, float b, float c, float *result1, float *result2);
+long double find_result(long double a, long double b, long double c, long double *result1, long double *result2);
+long double calculate_discriminant(long double a, long double b, long double c);
+long double find_linear_root(long double b, long double c, long double *result1);
+long double find_quadratic_roots(long double a, long double b, long double c, long double *result1, long double *result2);
 
-int find_result(float a, float b, float c, float *result1, float *result2) {
+long double find_result(long double a, long double b, long double c, long double *result1, long double *result2) {
     assert(result1 != NULL);
     assert(result2 != NULL);
 
@@ -20,11 +21,11 @@ int find_result(float a, float b, float c, float *result1, float *result2) {
     return find_quadratic_roots(a, b, c, result1, result2);
 }
 
-float calculate_discriminant(float a, float b, float c) {
+long double calculate_discriminant(long double a, long double b, long double c) {
     return b * b - 4 * a * c;
 }
 
-int find_linear_root(float b, float c, float *result1) {
+long double find_linear_root(long double b, long double c, long double *result1) {
     assert(result1 != NULL);
 
     if (!is_zero(b)) {
@@ -38,11 +39,11 @@ int find_linear_root(float b, float c, float *result1) {
     }
 }
 
-int find_quadratic_roots(float a, float b, float c, float *result1, float *result2) {
+long double find_quadratic_roots(long double a, long double b, long double c, long double *result1, long double *result2) {
     assert(result1 != NULL);
     assert(result2 != NULL);
 
-    float discriminant = calculate_discriminant(a, b, c);
+    long double discriminant = calculate_discriminant(a, b, c);
 
     if (!is_negative(discriminant)){
         if (is_zero(discriminant)) { 
@@ -50,8 +51,8 @@ int find_quadratic_roots(float a, float b, float c, float *result1, float *resul
             return kOneRoot;
 
         } else {
-            *result1 = (-b + sqrtf(discriminant)) / (2 * a);
-            *result2 = (-b - sqrtf(discriminant)) / (2 * a);
+            *result1 = (-b + sqrtl(discriminant)) / (2 * a);
+            *result2 = (-b - sqrtl(discriminant)) / (2 * a);
             sort_result(result1, result2);
             root_zero_checker(result1);
             root_zero_checker(result2);
